+++ qwilib/components/__init__.py
@@ -1,0 +1,3 @@
+from .displays import *
+from .forms import *
+from .inputs import *
