+++ qwilib/components/inputs.py
@@ -1,7 +1,11 @@
-from PySide6 import QtCore
-from PySide6 import QtWidgets
+__all__ = [
+    "SearchAndFilter",
+    "SearchAndCreate"
+]
 
-from qwilib import containers, inputs, displays, enums
+from PySide6 import QtCore, QtWidgets
+
+from qwilib import containers, inputs
 
 
 class SearchAndFilter(containers.HBoxContainer):
@@ -38,14 +42,3 @@ class SearchAndCreate(containers.HBoxContainer):
 
         self.add_widget(self.search_lineedit)
         self.add_widget(self.create_btn)
-
-
-class IconAndName(containers.HBoxContainer):
-    def __init__(self, name: str, icon_path: str = enums.Icons.DEFAULT_PROFILE, parent: QtWidgets.QWidget = None):
-        super().__init__(parent=parent)
-
-        icon = displays.SizedPixmapLabel(icon_path, parent=self)
-        name_display = displays.VCenterAlignedLabel(name, parent=self)
-        self.set_alignment(QtCore.Qt.AlignLeft)
-        self.add_widget(icon)
-        self.add_widget(name_display)
