+++ qwilib/inputs/__init__.py
@@ -1,0 +1,5 @@
+from .buttons import *
+from .calendars import *
+from .checkboxes import *
+from .comboboxes import *
+from .lineedits import *
