+++ qwilib/inputs/buttons.py
@@ -1,6 +1,26 @@
-from PySide6 import QtCore
-from PySide6 import QtGui
-from PySide6 import QtWidgets
+__all__ = [
+    "IconButton",
+    "RefreshButton",
+    "NotificationButton",
+    "FilterButton",
+    "CreateButton",
+    "CollapseButton",
+    "InfoButton",
+    "GridViewButton",
+    "ListViewButton",
+    "CancelButton",
+    "ApplyButton",
+    "ChangeButton",
+    "ImportTasksFromButton",
+    "ConnectParentsButton",
+    "ToggleSidebarButton",
+    "ToggleSearchContentsButton",
+    "ToggleColumnsButton",
+    "TogglePropertiesButton",
+    "ToggleThemeButton",
+]
+
+from PySide6 import QtGui, QtWidgets
 
 from qwilib import commons, enums
 
@@ -166,67 +186,3 @@ class ToggleThemeButton(QtWidgets.QPushButton):
         super().__init__(parent=parent)
         self.setText("Toggle Theme")
         self.setToolTip("Toggle the theme of the application")
-
-
-class NoScrollCombobox(QtWidgets.QComboBox):
-    def __init__(self, parent: QtWidgets.QWidget = None):
-        super().__init__(parent=parent)
-
-    def wheelEvent(self, event: QtGui.QWheelEvent):
-        pass
-
-
-class SearchLineEdit(QtWidgets.QLineEdit):
-    def __init__(self, parent: QtWidgets.QWidget = None):
-        super().__init__(parent=parent)
-        self.setPlaceholderText("Type here to search...")
-
-
-class ShowMyTaskCheckBox(QtWidgets.QCheckBox):
-    def __init__(self, parent: QtWidgets.QWidget = None):
-        super().__init__(parent=parent)
-        self.setText("Show My Task only")
-
-
-class CalendarDateEdit(QtWidgets.QDateEdit):
-    def __init__(self, parent: QtWidgets.QWidget):
-        super().__init__(parent=parent)
-
-        self.setDisplayFormat("d MMMM yyyy")  # E.g 18 August 2022
-        self.setDate(QtCore.QDate().currentDate())
-        self.setCalendarPopup(True)
-
-
-class FormWidget(QtWidgets.QWidget):
-    def __init__(self, label: str, parent: QtWidgets.QWidget = None):
-        super().__init__(parent=parent)
-        layout = QtWidgets.QFormLayout()
-        layout.setContentsMargins(0, 0, 0, 0)
-        self.setLayout(layout)
-
-        self.widget = self._define_widget()
-        label = label + ":"
-        layout.addRow(label, self.widget)
-
-    def _define_widget(self) -> QtWidgets.QWidget:
-        raise NotImplementedError("Override this method and return some widget.")
-
-
-class ComboBoxForm(FormWidget):
-    combobox_selected = QtCore.Signal(int)
-
-    def __init__(self, label: str, parent: QtWidgets.QWidget = None):
-        super().__init__(label, parent)
-        self.widget: QtWidgets.QComboBox = self.widget
-        self.widget.currentIndexChanged.connect(self.combobox_selected.emit)
-
-    def populate_combobox(self, items: list[str]):
-        self.widget.addItems(items)
-
-    def _define_widget(self) -> QtWidgets.QWidget:
-        return NoScrollCombobox()
-
-
-class ProjectComboBox(ComboBoxForm):
-    def __init__(self, parent: QtWidgets.QWidget = None):
-        super().__init__("Project", parent=parent)
