+++ qwilib/windows/__init__.py
@@ -1,0 +1,2 @@
+from .dialogs import *
+from .windows import *
